all global vars in one file
not sure if this is a good idea but w/e
